Making js instrumentation save to a single table
This is actually much less efficient in terms of space than the previous implementation.
Unfortunately, the previous indexing method doesn't work when there are multiple instances
multiple browsers. Our data aggregation doesn't allow access to return values of queries,
so there is simply no way to autoincrement an id, and return it for use as a foreign key
in another table.