Reworking browser restart pipline to fix crashes and deadlocks. In support of #19
Rather than having multiple threads managing the same browser, each
thread will simply set a flag if the browser needs to be restarted.
Then, the main thread managing the browser will restart if necessary.
This fixes a bunch of race conditions caused by managing the process in
two separate threads. Also bumped the failure threshold as it was too
low for instances with a low number of browsers.