Bugfix: JSON serialize headers, fixing cookie parsing.
Mitmproxy changed the way it stores headers to use
netlib.odict.ODictCaseless. This type has a string representation of a
list of lists that must be eval()'ed to parse the headers. Instead,
let's use json serialization in combination with the save/load functions
provided by the ODictCaseless class. This fixes the bug that would cause
http cookie tables to be empty with newer versions of mitmproxy.