Save javascript in LevelDB database instead of flat files
Flat files are fine for smaller crawls, however the structure begins to
break down for long crawls. Directories with millions of small files are
slow to interact with, especially on traditional hard disks. LevelDB
requires a few extra dependencies.

A third (unrelated) dependency is added to fix the install of Pillow.