Fixing bug in HTTP Cookie parsing that skipped some Set-Cookies.
The bug arose from the fact that header_id is not unique, both
http_request and http_response headers start at 0. When we checked to
see if the header was already parsed we checked if it's header ID was
present in the http_cookies database, but forgot to include the
http_type in that search. This has been a common mistake during analysis
as well. For this reason, I'm splitting the http_cookies table into
http_request_cookies and http_response_cookies, as the cookies are
conceptually different despite http_request cookies having a subset of
http_response Set-Cookie fields (i.e. name and value).