Better Handling of Child Process Exceptions
Child processes can now raise their exceptions to the TaskManager, which
both gives the child processes the ability to stop the crawl (for
critical exceptions), and gives the TaskManager the ability to cleanup
before crashing. In doing so, I've largely standarized the
BrowserManager <--> {TaskManager, Browser} pipe. I also added in
ProfileLoadError as an example of a child --> parent exception.