Fixing encoding issues in extension.
I've added UTF-8 encoding to the escape function and made sure the
function is being called for any field that may contain non-ascii
characters. Thus, we should no longer have to deal with encoding issues
on the receiving end of the socket. Encoding utf-8 -> unicode can happen
elsewhere in the application. As it stands, sqlite is fine writing utf-8
encoded values for these queries.