Adding more flexible handling of exceptions in the child process.
It's now possible to raise exceptions during command execution that will
reraise in the main process. I've currently added `AssertionError` as
an exception which kills the main process, which should make it easy to
write tests with assertions in the child process (e.g. assert depending
on the state of a page). This required reworking the way the command
execution threads handle exceptions.