Avoid storing non-headers from POST messages in the headers column.
We were storing all the request headers from upload stream in the
headers column. We found out that in some cases,
non-header data are interpreted as headers since they have the same
key:value format.

With this change, we'll store the headers selectively.

Remove the handling of fallback to POST "lines" when the post body is
empty. We haven't observed this fallback on the 6000 sites we
crawled (using the record_post branch).