Adding testing platform configuration flag.
A `testing` boolean is available within `manager_params` in the python
portion of the platform, in the `config` dictionary in the extension
scripts, and in the `testing` variable for the injected content script.
The flag is being added to make it possible to expose different
functionality throughout the platform when it is being tested, whether
manually or in automated tests.

As an example, we expose the `instrumentObject` method on the page
global object `window` when testing.