Refactor instrumentObject so functions are fully instrumented.
Previously function instrumentation was treated as an alternative operation
to other property instrumentation. This isn't necessary as functions
are accessed the same way other object properties are, so we should
instrument access to them in the same way. Instead, the logging of a
function calls is done by additional instrumentation of the returned
function.