Further workaround for Python 3 + Selenium 3 + FIFOs.
The behavior of `open(path, "a+")` differs between Python 2 and Python 3.
In the latter, it will try to seek to the end, and if this fails (e.g. if
`path` is a pipe) it will throw an exception.  To work around this we
have to monkey-patch selenium.webdriver.firefox.service.Service.