Bugfix in HTTP Instrument + fixing tests [wip]
Most of the HTTP Instrument tests broke due to a new column being added
in the middle of the http request and response tables. Our tests
shouldn't have such a strong dependence on column order. Let's move
query responses over to the `sqlite3.Row` datatype, which allows access
by name. This is work in progress, I expect a bunch of other tests to
fail on calls to `db_utils::query_db`.