Provide JavaScript 1st/3rd party information
By adding the fields document_url, top_document_url, and referrer_url to the javascript table, it is now possible to determine whether a JavaScript was executed in a 1st or 3rd party context. The easiest way to do that is by checking whether top_document_url is NULL, which indicates that this script is a third party JavaScript. Other evaluations are possible too.