Move timestamp to the content script part
This patch moves the generation of the timestamp for the message from the page script part to the content script part. The page script code is execute with all the other JavaScript code of the website, and when a website decides to change the Date prototype to something else, then the previous variant might generate incorrect timestamps. By generating the timestamp in the content script part, the timestamp is not affected by any JavaScript code running in the website.