Add partitioning by instance + make visit_id 64 bits.
The intention of adding partitioning by instance_id is that it will be
easier to read slices of a large dataset (e.g. the first 100k sites out
of a million site dataset). An instance_id partitioning doesn't add any
overhead as different instances will already store data in different
parquet files. To help with slicing, I've also added a site_visits
index, which maps the instance id to everything stored in the
site_visits table.