Fix the Dockerfile
For some reasons, the build-extension.sh script does not work as root. This can be fixed by creating an openwpm user with sudo permissions and run the install.sh script as openwpm.

Currently, the openwpm user remains in the image, but the Dockerfile could also be changed to remove the user after the build again.