Use base64 encoding for the response body
The response body might contain binary content that cannot be correctly represented by a string. This patch assumes that the openwpm-webext-instrumentation returns the response body as an Uint8Array instead of a string. It is then encoded using base64 to make it string-safe and then send to the DataAggregator of OpenWPM to be saved in a database there.

This patch needs a change in openwpm-webext-instrumentation which is currently not available in the official repository. As long as mozilla/openwpm-webext-instrumentation#45 has not been merged, the patch will not work correctly.